V.1.0. First commit. Installing package -> MUI. Deleting unused files.
